requirements.txt

fastapi==0.112.2
uvicorn==0.30.6
streamlit==1.38.0
torch==2.3.1
Pillow==10.4.0
requests==2.32.3
ultralytics==8.2.86
